Improve format tests
